# Optimise squared, Mul, and Exp