```
name                     old time/op  new time/op  delta
Mul/single/uint256-8     8.23ns ± 2%  6.90ns ± 1%  -16.11%  (p=0.000 n=10+9)
Mul/single/big-8         64.5ns ± 1%  62.8ns ± 1%   -2.65%  (p=0.000 n=9+9)
_Exp/large/big-8         23.7µs ± 1%  23.7µs ± 1%     ~     (p=0.896 n=10+10)
_Exp/large/uint256-8     3.34µs ± 1%  2.72µs ± 3%  -18.63%  (p=0.000 n=8+10)
_Exp/small/big-8         7.12µs ± 2%  7.09µs ± 0%     ~     (p=0.345 n=10+9)
_Exp/small/uint256-8      283ns ± 1%   229ns ± 1%  -19.22%  (p=0.000 n=9+10)
Square/single/uint256-8  6.27ns ± 1%  5.14ns ± 1%  -17.91%  (p=0.000 n=10+10)
Square/single/big-8      60.0ns ± 2%  60.0ns ± 1%     ~     (p=0.781 n=10+10)
```