@chfast 2020-06-04
My benchmarks:
```
name                     old time/op  new time/op  delta
Mul/single/uint256-8     7.95ns ± 0%  6.56ns ± 0%  -17.46%  (p=0.000 n=16+18)
Mul/single/big-8         59.4ns ± 1%  59.2ns ± 0%     ~     (p=0.274 n=19+17)
Square/single/uint256-8  5.76ns ± 0%  4.94ns ± 0%  -14.20%  (p=0.000 n=18+19)
Square/single/big-8      57.5ns ± 0%  57.5ns ± 0%     ~     (p=0.077 n=18+18)
_Exp/large/big-8         21.3µs ± 1%  21.3µs ± 0%     ~     (p=0.925 n=20+18)
_Exp/large/uint256-8     3.00µs ± 0%  2.54µs ± 0%  -15.61%  (p=0.000 n=18+16)
_Exp/small/big-8         6.33µs ± 1%  6.34µs ± 1%     ~     (p=0.361 n=19+20)
_Exp/small/uint256-8      254ns ± 1%   217ns ± 0%  -14.42%  (p=0.000 n=20+15)
```