@chfast 2020-06-04
The change you reverted in `umul()` is also worth having.
```
name                     old time/op  new time/op  delta
MulMod/small/uint256-8   33.6ns ± 0%  32.3ns ± 0%  -3.97%  (p=0.000 n=19+17)
MulMod/mod64/uint256-8   97.5ns ± 1%  96.1ns ± 0%  -1.48%  (p=0.000 n=20+18)
MulMod/mod128/uint256-8   176ns ± 0%   171ns ± 0%  -2.84%  (p=0.000 n=17+17)
MulMod/mod192/uint256-8   173ns ± 0%   168ns ± 0%  -3.15%  (p=0.000 n=20+20)
MulMod/mod256/uint256-8   169ns ± 0%   162ns ± 0%  -4.14%  (p=0.000 n=17+18)
```