@holiman 2020-06-05
@yperbasis try a `git checkout fast_mul && git fetch master && git rebase -i origin/master` , which remove empty merge-commits.  

And then followed by `git push -f` if all went well.