@yperbasis 2020-06-05
> @yperbasis try a `git checkout fast_mul && git fetch origin && git rebase -i origin/master` , which remove empty merge-commits.
> (edited, should be `git fetch origin`)
> And then followed by `git push -f` if all went well.

Done, but some history appears to still linger for some reason.