@yperbasis 2020-06-05
> (and add back the improvement to `umul`)

Done