@yperbasis 2020-06-05
> Ok, don't worry about it - we can always squash-merge.
> 
> However, If you do want it explicitly it to be commit-merged, then it needs to be rebased, empty commits removed and/or squashed to relevant commits.

I'm fine if you squash-merge it.